them anh tra ve va update cho review

+++ LVTN_BE_COFFE/Domain/VModel/ReViewVModel.cs
@@ -10,7 +10,7 @@ namespace LVTN_BE_COFFE.Domain.VModel
         [Required(ErrorMessage = "Điểm đánh giá là bắt buộc.")]
         // Thêm ràng buộc để đảm bảo điểm nằm trong phạm vi cho phép (ví dụ: 1 đến 5)
         [Range(1, 5, ErrorMessage = "Điểm đánh giá phải từ 1 đến 5.")]
-        public int Rating { get; set; }
+        public float Rating { get; set; }
 
         [StringLength(500, ErrorMessage = "Bình luận không được vượt quá 500 ký tự.")]
         public string? Comment { get; set; }
@@ -22,7 +22,7 @@ namespace LVTN_BE_COFFE.Domain.VModel
 
         [Required(ErrorMessage = "Điểm đánh giá là bắt buộc.")]
         [Range(1, 5, ErrorMessage = "Điểm đánh giá phải từ 1 đến 5.")]
-        public int Rating { get; set; }
+        public float Rating { get; set; }
 
         [StringLength(500, ErrorMessage = "Bình luận không được vượt quá 500 ký tự.")]
         public string? Comment { get; set; }
